test: 테스트 케이스 병합
